dataset_viewer/requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
zarr>=2.14.0
numpy>=1.24.0
streamlit-drawable-canvas-fix>=0.9.0
